v2.4.3.6 -> v2.4.3.7
  - Johannes Erdfelt: USB updates
  - David Howells: more rw-sem stuff
  - David Miller: network callback cleanups and fixes
  - Jan Harkes: make Coda use the proper VFS layer interfaces, so that it can use
  "non-traditional-unix" filesystems without inode numbers for backing store.

+++ fs/affs/namei.c
@@ -226,7 +226,7 @@ affs_lookup(struct inode *dir, struct dentry *dentry)
 		u32 ino = bh->b_blocknr;
 
 		/* store the real header ino in d_fsdata for faster lookups */
-		dentry->d_fsdata = (void *)ino;
+		dentry->d_fsdata = (void *)(long)ino;
 		switch (be32_to_cpu(AFFS_TAIL(sb, bh)->stype)) {
 		case ST_LINKDIR:
 		case ST_LINKFILE:
